Web site updates across multiple pages
Home page (readme.md): Added a prominent tip about our LFEnergy presentation.
Participants: Just added anchors to the names to link from other pages.
Meetings: Added a section on the LFEnergy Presentation to give it a permanent home. Reversed the order of the meetings to be most recent first.
References:
1 Added a section of Microgrid vendors that links to the various companies we have dealt with including actual grid vendors like Mesh Power and Solshare. Include Libre.Solar so you might want to edit.
2. Added a new section on Embedded Systems, Tools and Firmware to capture
links to the various tools and things. This might move to the Implementation section eventually but wanted a place that I could reference in emails etc.
3. Added a few new links in various other sections.
Added a few new links in various sections.